changed folder structure of frontend for good
