+++ .gitignore
@@ -18,6 +18,7 @@ backend/node_modules/
 .env.development.local
 .env.test.local
 .env.production.local
+.env
 
 npm-debug.log*
 yarn-debug.log*
