+++ package.json
@@ -59,6 +59,7 @@
   },
   "devDependencies": {
     "@hookform/devtools": "^4.3.1",
+    "dotenv": "^16.3.1",
     "react-app-rewired": "^2.2.1"
   }
 }
