+++ src/API/MyDocument/MyDocument.jsx
@@ -11,7 +11,7 @@ import {
 //register fonts
 
 import React from "react";
-import logo from '../images/logo.png';
+import logo from './logo.png';
