+++ src/Components/smallComponents/LoginOptions.jsx
@@ -6,14 +6,11 @@ import Avatar from '@mui/material/Avatar';
 import Menu from '@mui/material/Menu';
 import Typography from '@mui/material/Typography';
 import MenuItem from '@mui/material/MenuItem';
-import { UserContext } from '../../App';
-import {useContext} from 'react';
 import { useNavigate } from 'react-router-dom';
 const settings = ['Profile', 'Account', 'Dashboard', 'Logout'];
 
 export default function LoginOptions(){
 
-  const [isLoggedIn, setLogin] = useContext(UserContext);
   const Navigate = useNavigate();
     const [anchorElUser, setAnchorElUser] = useState(null);
     const handleOpenUserMenu = (event) => {
@@ -53,7 +50,6 @@ export default function LoginOptions(){
                    setAnchorElUser(null);
                     if(setting==='Logout'){
                       Navigate('/');
-                      setLogin(false);
                     }
             }} name={setting}>
               <Typography  textAlign="center">{setting}</Typography>
