--- src/Database.jsx
@@ -1,6 +1,0 @@
-const Admin ={
-    username:'admin',
-    password:'admin'
-}
-
-export default Admin;
